fix: shift all Helpers to UI chore: ignore prettier on mdx files fix: rename components and stories
