+++ src/components/AssetInfo/TokenSideBarTooltip.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactTooltip from "react-tooltip";
-import { SvgIcon } from "./../Helpers/SvgIcon";
+import { SvgIcon } from "./../UI/SvgIcon";
 
 interface TokenSideBarTooltipProps {
   id: string;
