--- src/components/TitleTransferPanel/Address/Address.stories.mdx
@@ -1,43 +1,0 @@
-import { Meta, Story, Props, Preview } from "@storybook/addon-docs/blocks";
-import { Address } from "./Address";
-import { ExternalLink } from "./../../Helpers/ExternalLink";
-
-<Meta title="TitleTransferPanel/Address" component={Address} />
-
-# Address
-
-Address with information.
-
----
-
-### Props
-
-<Props of={Address} />
-
----
-
-### Address
-
-With title, name and address.
-
-<Preview>
-  <Story name="Default">
-    <Address title="Role" name="Name (Identifier)">
-      <div>0x1D350495B4C2a51fBf1c9DEDadEAb288250C703e</div>
-    </Address>
-  </Story>
-</Preview>
-
----
-
-### Address (with link)
-
-With title, name and address link.
-
-<Preview>
-  <Story name="AddressExternalLink">
-    <Address title="Role" name="Name (Identifier)">
-      <ExternalLink name="0x1D350495B4C2a51fBf1c9DEDadEAb288250C703e" href="#" />
-    </Address>
-  </Story>
-</Preview>
