--- src/components/TitleTransferPanel/Address/index.tsx
@@ -1,1 +1,0 @@
-export * from "./Address";
