+++ src/components/TitleTransferPanel/AddressInfo/AddressInfo.module.scss
@@ -1,6 +1,6 @@
 @import "../../../styles/abstracts/__abstracts.scss";
 
-.address {
+.address-info {
   h5,
   h6 {
     @include font-sourcesanspro-bold;
