+++ src/components/TitleTransferPanel/AddressInfo/AddressInfo.stories.mdx
@@ -1,0 +1,61 @@
+import { Meta, Story, Props, Preview } from "@storybook/addon-docs/blocks";
+import { AddressInfo } from "./AddressInfo";
+import { ExternalLink } from "./../../UI/ExternalLink";
+
+<Meta title="TitleTransferPanel/AddressInfo" component={AddressInfo} />
+
+# AddressInfo
+
+Address with information.
+
+---
+
+### Props
+
+<Props of={AddressInfo} />
+
+---
+
+### AddressInfo
+
+With title, name and text address.
+
+<Preview>
+  <Story name="Default">
+    <AddressInfo title="Title" name="Name (Identifier)">
+      <>0x1D350495B4C2a51fBf1c9DEDadEAb288250C703e</>
+    </AddressInfo>
+  </Story>
+</Preview>
+
+---
+
+### AddressInfo (Holder)
+
+With actual title role, name and external link address.
+
+<Preview>
+  <Story name="AddressInfoHolder">
+    <AddressInfo title="Holder" name="Bank of China">
+      <ExternalLink name="0x1D350495B4C2a51fBf1c9DEDadEAb288250C703e" href="#" />
+    </AddressInfo>
+  </Story>
+</Preview>
+
+---
+
+### AddressInfo (BL info)
+
+Bill of Lading information.
+
+<Preview>
+  <Story name="AddressInfoBL">
+    <AddressInfo title="BL information">
+      <>
+        <ExternalLink name="View BL Registry" href="#" />
+        <br />
+        <ExternalLink name="View Endorsement Chain" href="#" />
+      </>
+    </AddressInfo>
+  </Story>
+</Preview>
