+++ src/components/TitleTransferPanel/AddressInfo/index.tsx
@@ -1,0 +1,1 @@
+export * from "./AddressInfo";
