+++ src/components/TitleTransferPanel/TitleView/BLinfo.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { useSelector } from "react-redux";
 import css from "./TitleView.module.scss";
-import { Address } from "../Address";
-import { ExternalLink } from "../../Helpers/ExternalLink";
+import { AddressInfo } from "../AddressInfo";
+import { ExternalLink } from "../../UI/ExternalLink";
 import { getAssetInfo } from "../../../utils";
 import { makeEtherscanAddressURL, makeEtherscanTokenURL } from "../../../utils";
 
@@ -16,13 +16,13 @@ export const BLinfo = () => {
 
   return (
     <div className={css["title-view"]}>
-      <Address title="BL information">
+      <AddressInfo title="BL information">
         <>
           <ExternalLink name="View BL Registry" href={tokenRegistryHref} />
           <br />
           <ExternalLink name="View Endorsement Chain" href={tokenHistoryHref} />
         </>
-      </Address>
+      </AddressInfo>
     </div>
   );
 };
