--- src/components/TitleTransferPanel/TitleView/Beneficiary.stories.mdx
@@ -1,26 +1,0 @@
-import { Meta, Story, Props, Preview } from "@storybook/addon-docs/blocks";
-import { Beneficiary } from "./Beneficiary";
-
-<Meta title="TitleTransferPanel/TitleView" component={Beneficiary} />
-
-# TitleView (Beneficiary)
-
-TitleView.
-
----
-
-### Props
-
-<Props of={Beneficiary} />
-
----
-
-### Beneficiary
-
-TitleView of BeneficiaryAddress.
-
-<Preview>
-  <Story name="Beneficiary">
-    <Beneficiary beneficiaryAddress="0x1D350495B4C2a51fBf1c9DEDadEAb288250C703e" />
-  </Story>
-</Preview>
