--- src/components/TitleTransferPanel/TitleView/Holder.stories.mdx
@@ -1,26 +1,0 @@
-import { Meta, Story, Props, Preview } from "@storybook/addon-docs/blocks";
-import { Holder } from "./Holder";
-
-<Meta title="TitleTransferPanel/TitleView" component={Holder} />
-
-# TitleView (Holder)
-
-TitleView.
-
----
-
-### Props
-
-<Props of={Holder} />
-
----
-
-### Holder
-
-TitleView of holderAddress.
-
-<Preview>
-  <Story name="Holder">
-    <Holder holderAddress="0x1D350495B4C2a51fBf1c9DEDadEAb288250C703e" />
-  </Story>
-</Preview>
