+++ src/components/UI/Button/Button.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, Story, Props, Preview } from "@storybook/addon-docs/blocks";
 // import { action } from '@storybook/addon-actions';
 import { ButtonSolid, ButtonIcon, ButtonCircle } from "./Button";
-import { SvgIcon, SvgIconEdit, SvgIconBook, SvgIconX } from "./../../Helpers/SvgIcon";
+import { SvgIcon, SvgIconEdit, SvgIconBook, SvgIconX } from "./../../UI/SvgIcon";
 
 <Meta title="UI/Button" component={ButtonSolid} />
