+++ src/components/UI/SvgIcon/SvgIcon.test.tsx
